Apply bidirectional queue design to the input system
This commit splits the input system into two parts, event and submission, and
introduces two system calls, setup_queue and submit, the former is used to
assist in the creation of queues, which is part of the communication bridge
between the user application code and the emulator, and the latter is used to
notify the emulator that a submission (or a command) should be processed. The
design of these system calls is heavily based on Linux io_uring interface.