Migrate to RISC-V Compatibility Framework (RISCOF)
This commit updates the version of riscv-arch-test and introduces the
RISC-V Architectural Test Framework (RISCOF) into emulator.

In riscv-arch-test part, the following shows the old and revised version.
- Old: branch old-framework-2.x
- New: branch main and commit 933dddd

In RISCOF part, it uses config.ini to configure the plugins to be used —
the device-under-test (DUT) and the reference model (REF). The ISA
specification defines the available ISA extensions in emulator and users
can modify via YAML file in DUT's folder. Besides, the file setup.py generates
 the config.ini automatically and modifies the ISA specification based on user's
input.
- DUT: rv32emu
- REF: sail_cSim

In the other hand, RISCOF uses riscv-gnu-toolchain to compile the tests in
riscv-arch-test. Therefore, change the toolchain from riscv-none-elf-gcc-xpack
to riscv-gnu-toolchain in Continuous Integration (CI).

Finally, summarize the test report based on the latest revision of rv32emu.
- Passed ISA in riscv-arch-test: I, M, C, Zicsr (privilege), Zifencei

Close #49