Implement LFU as default cache along with memory pool (#125)
An adaptive replacement cache (ARC) was implemented to reduce memory usage.
However, the complexity of the replacement policy of ARC can lead to a
decrease in overall performance. Therefore, a least frequently used (LFU) cache
was also implemented, which outperformed ARC in testing. However, the ARC was
retained for future use.

In addition, a memory pool was imported to limit the memory usage of both caches.

Performance statistics for the LFU cache and ARC running the CoreMark benchmark
indicate that the LFU cache outperforms ARC.

| Test |  CoreMark(Iterations/Sec)  |
|------+----------------------------|
| ARC  |         1123.776           |
| LFU  |         1155.174           |