Link the register jump (#224)
The interpreter exited the execution loop when encountering a register
jump instruction, resulting in a significant performance problem. To
tackle this issue, the solution involves searching for the subsequent
block at the conclusion of a register jump's execution. Upon locating
the next block, the register jump is linked to it.

As shown in below analysis, this modification improves the overall
performance by 1~2%.

* Core i7-11700

| Metric   |    original     |    proposed    | speedup |
|----------+-----------------+----------------+---------|
| CoreMark | 1931.047 iter/s | 1956.64 iter/s | +1.33%  |
| stream   |   79.156 sec    |   77.123 sec   | +2.57%  |

Close #223