Refine clang-format style
AttributeMacros: A vector of strings that should be interpreted as
attributes/qualifiers instead of identifiers.

StatementAttributeLikeMacros: Macros which are ignored in front of
a statement, as if they were an attribute.

StatementMacros: Typical macros are expressions, and require a
semi-colon to be added; sometimes this is not the case, and this
allows to make clang-format aware of such cases.