Drop memory_read_str function
The memory_read_str function tends to terminate the string with '\0'
character but the string is only terminated with '\0' character if src is
less than max characters long according to "man 3 strncpy". As a result,
there is a chance to make the dst is not terminated with '\0' character,
e.g., the length of filename equals 256 characters long. It can simply
solved by passing "sizeof(name_str) - 1" to memory_read_str function
instead of "sizeof(name_str)" when calling memory_read_str function.
But, to deal with the filename >= 256 characters long, malloc the desired
length of filename during runtime is a better way.