Add code cache flush mechanism
This modification introduces a code cache flush mechanism to handle the
situation when the code cache becomes full. We flush the code cache by
setting the offset to its initial value and clearing all the recording
data for the code cache. Subsequently, we restart the T1C translation
for the target chained-block.

Close: #365