Raise illegal instruction exception for Zicsr
According to the RISC-V privileged architecture 20211203 (Section 2.1),
write a read-only register should raise illegal instruction exception.
Also, unprivileged ISA 20191213 states that a CSR instruction is
considered a write when the rs1 or uimm is not zero (check Table 9.1).
Since the exceptions are raised during the decoding stage, the execution
of csr_csrrw, csr_csrrs, and csr_csrrc function can skip checking the
writability of csr anymore.