Optimize syscall_write() to avoid redundant checks and FILE* retrievals
There is redundant checking of map_at_end within a loop, along with
repetitive calls to map_iter_value() to retrieve the file pointer. This
results in unnecessary overhead.

Address the issue by optimizing the code to perform the map_at_end
check and retrieve the file pointer only once at the beginning.