CI: Reduce HTTP requests
Check prebuilt tarballs are existing or not first. If so, skip the
downloading of releases tags and tarballs.

In the CI, we first call "make artifact" to prepare the tarballs, so the
GitHub API for fetching releases tags would not be called and be skipped
when "make" is called next time.

After this commit, wget in the CI would not complain about error code
403 (rate limit exceeded) anymore. However, the build on Aarch64 still
failed because of other reansons, e.g. software bugs of GCC or QEMU.

Close #536