Allow emcc build
Emcc generates Wasm but not a native executable, making the generated
bin2c non-executable. As a result, the device tree blob cannot be built
into a C application. To resolve this, fall back to a native compiler
gcc or clang to build bin2c.

Additionally, ensuring CC is set to gcc, clang, or emcc. Otherwise,
trigger a build error early.

Related: #534