Support readonly feature of VirtIO block device
Add support for comma-separated options for VirtIO devices. For the
readonly option, if the string readonly is specified, subsequent VirtIO
block device access will be read-only; otherwise, the default behavior
is read and write. To implement this, a device_features field is
introduced to store device features during initialization.
VIRTIO_BLK_F_RO is the feature bit to enable readonly.

This feature enhances the security of the VirtIO block device by
preventing unintended data corruption in the guest OS when accessing
critical data.

Additionally, after implementing the VirtIO block device in #539, some
register access errors were shown and they are fixed in this commit:
- reading from the VirtIO DeviceFeatures register now correctly checks
  the device_features_sel instead of driver_features_sel.
- writing to the VirtIO DeviceFeaturesSel register now correctly updates
  device_features_sel instead of driver_features_sel.