Dynamically configure VirtIO node of Device Tree
PR #584 initializes the vblk only when the vblk CLI option is provided.
This leads to a segmentation fault when vblk is not specified, as the
capacity of the vblk device is left uninitialized (not set to 0).
Previously, the code would always initialize vblk but set its capacity
to 0, preventing the VirtIO block driver from interacting with it and
avoiding the segmentation fault. However, this original approach was
functionally redundant.

A better solution is to dynamically load or unload the VirtIO node in
the SoC node of the Device Tree based on configuration. To support
future others VirtIO devices, the interface of load_dtb() has been
refined, it now can parse all options stored in the attribute.